+++ src/Join-Until_function.ps1
@@ -4,6 +4,10 @@
 
     Concatenate lines until they match the specified regular expression.
 
+    Param:
+      -TrimOnlyInJoin ... Trim only the connected lines before joining.
+      -DisableTrim    ... Disable Trim all lines before and after joining.
+
 .LINK
     Join-While, Join-Until, Trim-EmptyLine, list2txt, csv2txt
 
@@ -76,7 +80,7 @@
 
         # data
         2024-10-24
-        <blankline>
+
         btklab, fuga
         summary-1 summary-2 summary-3
         link-1 link-2
@@ -134,43 +138,44 @@ function Join-Until {
             Mandatory=$False,
             Position = 0
         )]
-        [Alias('m')]
+        [Alias('r')]
         [String] $Regex = "."
         ,
         [Parameter(
             HelpMessage="Output delimiter",
-            Mandatory=$False,
-            Position = 1
+            Mandatory=$False
         )]
         [Alias('d')]
         [String] $Delimiter = " "
         ,
         [Parameter(
-            HelpMessage="Do not trim line",
+            HelpMessage="Trim only injoin line",
             Mandatory=$False
         )]
-        [Alias('nt')]
-        [Switch] $NoTrim
+        [Switch] $TrimOnlyInJoin
         ,
         [Parameter(
-            HelpMessage="Do not skip blank line",
+            HelpMessage="Disable Trim line",
             Mandatory=$False
         )]
-        [Alias('ns')]
-        [Switch] $NoSkip
+        [Switch] $DisableTrim
+        ,
+        [Parameter(
+            Mandatory=$False
+        )]
+        [Switch] $DisableBlankDelimiter
         ,
         [Parameter(
             HelpMessage="Skip header until a specified string is found",
             Mandatory=$False
         )]
-        [Alias('s')]
         [Switch] $SkipHeader
         ,
         [Parameter(
             HelpMessage="Delete matched strings",
             Mandatory=$False
         )]
-        [Alias('dm')]
+        [Alias('del')]
         [Switch] $Delete
         ,
         [Parameter(
@@ -184,12 +189,14 @@ function Join-Until {
             HelpMessage="Insert Before",
             Mandatory=$False
         )]
+        [Alias('b')]
         [String[]] $Before
         ,
         [Parameter(
             HelpMessage="Insert After",
             Mandatory=$False
         )]
+        [Alias('a')]
         [String[]] $After
         ,
         [parameter(
@@ -204,6 +211,7 @@ function Join-Until {
         # set variables
         [Int] $rowCounter = 0
         [Bool] $firstMatch = $False
+        [Bool] $inJoin = $True
         [String[]] $tempLineAry = @()
         $tempAryList = New-Object 'System.Collections.Generic.List[System.String]'
         ## private function
@@ -221,23 +229,27 @@ function Join-Until {
     process {
         $rowCounter++
         [String] $readLine = [String] $_
-        if ( $NoTrim ){
-            #pass
-        } else {
-            [String] $readLine = $readLine.Trim()
-        }
         if ( $readLine -match $Regex ){
             [Bool] $firstMatch = $True
+            [Bool] $inJoin = $False
+        } else {
+            [Bool] $inJoin = $True
+        }
+        if ( -not $DisableTrim ){
+            [String] $readLine = $readLine.Trim()
+        } elseif ( $TrimOnlyInJoin -and $inJoin ){
+            [String] $readLine = $readLine.Trim()
         }
         ## test readline
         Write-Debug "firstMatch: $firstMatch"
         if ( $SkipHeader -and -not $firstMatch ){
             return
         }
-        if ( $NoSkip ){
-            #pass
-        } else {
-            if ( $readLine -eq ''){
+        if ( $readLine -match '^\s*$'){
+            if ( $DisableBlankDelimiter ){
+                return
+            } else {
+                #Write-Output ''
                 return
             }
         }
@@ -246,10 +258,8 @@ function Join-Until {
             [String[]] $tempLineAry = $tempAryList.ToArray()
             if ( $tempLineAry.Count -gt 0 ){
                 [String] $writeLine = $tempLineAry -join $Delimiter
-                if ( $NoTrim ){
-                    #pass
-                } else {
-                    [String] $writeLine = $writeLine.Trim()
+                if ( -not $DisableTrim ){
+                    $writeLine = $writeLine.Trim()
                 }
                 Write-BeforeAndAfterOutput $writeLine
             }
@@ -263,11 +273,6 @@ function Join-Until {
         } elseif ( $Delete ){
             [String] $readLine = $readLine -replace $Regex, ''
         }
-        if ( $NoTrim ){
-            #pass
-        } else {
-            [String] $readLine = $readLine.Trim()
-        }
         $tempAryList.Add( $readLine )
     }
 
@@ -276,10 +281,8 @@ function Join-Until {
         [String[]] $tempLineAry = $tempAryList.ToArray()
         if ( $tempLineAry.Count -gt 0 ){
             [String] $writeLine = $tempLineAry -join $Delimiter
-            if ( $NoTrim ){
-                #pass
-            } else {
-                [String] $writeLine = $writeLine.Trim()
+            if ( -not $DisableTrim ){
+                $writeLine = $writeLine.Trim()
             }
             Write-BeforeAndAfterOutput $writeLine
         }