+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
-- NA
+- Changed [Join-While][] Option `Trim` to `DisableTrim`
+- Changed [Join-Until][] Option `Trim` to `DisableTrim`
 
 ## [0.5.0] - 2025-03-26